A start at a dLua spells library.
Provides wrappers for non-player Toxic Radiance and Ozocubu's
Refrigeration. These are basically just an adjustment of kill categories
and messages to allow them to be used as background effects in WizLabs
(Ozocubu's and Olgreb's).

Hopefully a wrapper for monster (and possibly player)-cast bolt
structures can also be included, which could then be accessed with the
(planned) Lua traps functionality.