Change some colours and glyphs as per FR#250.
* Queen bees become yellow 'k'.
* Bees become brown 'k'.
* Queen ants become lightred 'a'.
* Minotaurs become lightred 'H'.
* Trapdoor spiders become lightcyan 's'.
* Wolf spiders become white 's'.
* Giant cockroaches become brown 's'.

Of all of the changes, the most disconcerting may be the minotaur and
the cockroach; the minotaur is not often seen, and as it is actually a
hybrid, is more properly represented on the 'H' glyph.

The giant cockroach is seen regularly throughout the early game, and it
may cause confusion to old players who are used to brown 's' being
either trapdoor spiders of wolf spiders. However, there are currently
more insects on the 's' glyph than there are on the 'a' glyph, which
itself is described as being dedicated to insects.

As queen bees have been moved onto 'a', it makes more sense for them to
be on their own on this glyph, similarly to the fact that bees are on
their own on the 'k' glyph.

This now means that the following glyphs are empty: 'i', 'Q', 't', '6'.