Colour-variant "white smoke" tile.
Basically just a test to make sure that cloud recolouring using
recoloured tiles (if they exist) works. As far as I can see, it works
exactly as hoped-for! It's just a matter of reorganising the base cloud
associations so that you can specify any type of smoke, recolour it, and
get the proper tile.