Do not destroy items in deep water and lava
In the past, this made sense, because deep water was forever.  Now,
with Shoals tides and Fedhas Sunlight, deep water can dry up, and it
makes no sense for waterproof items to cease to exist after a stint
as a reef.  Scrolls are still destroyed in lava because they are
flammable.