Add wave tiles to smoothen the transition between shallow water and floor.
Currently only applies to Shoals, partly because the relevant base tiles are
only used in the Shoals, partly because it's experimental. The checks needed
for adjacent water tiles increase the drawing time, but I'm not sure by how
much, seeing how in my test runs I only measured the seconds used for &O and
came up with no noticeable difference. Still might increase lags on computers
already prone to slow tiles, in which case we'll need to revert this change.

Making the dungeon look less "gridlike" greatly improves the graphics and
makes me want to extend the concept to floor/floor transitions, but it could
potentially harm the interface. Not sure where to draw the line, and in any
case the lag appears to be the more imminent problem.

I've also played around with also adding waves for the shallow/deep water
transitions but this would require special behaviour also for the water masks
used for monsters and items, which is not a problem for floor tiles.