Remove hand crossbows.
This involves a lot of changes, and also increments TAG_MAJOR_VERSION to
13. The unrand "Sniper" is now a crossbow (hopefully its current title
is still suitable for this task), and thieves no longer start with hand
crossbows (but still retain darts).

Wanderers with crossbow skills will now get a crossbow and bolts. This
may be overpowered, but I didn't want to change the code here too much.
It can always be adjusted.

This removes the "quiver_type" enum from enum.h, and uses NUM_AMMO in
the player.h quiver vector instead. Random elves that generated with
hand crossbows before no longer do so, and this has not been replaced by
normal crossbows.

I hope that's everything!