Subsume Darts skill into Throwing.
Instances where Darts skill was previously check (such as blow guns,
throwing of darts, etc), now check the throwing skill instead. This
causes a few problems:

* missile acquirement is possibly skewed now
* aptitudes are now wrong, and should be adjusted

This doesn't include code to compensate for large monsters that used to
have a good throwing skills and bad dart skills now using the lower
throwing skill, and vice versa, but it's a major start.

TAG_MAJOR_VERSION was bumped in the previous commit. Hopefully I haven't
broken anything too major, but if so, the commit can be reverted.