Continuation of the range combat overhaul: branding spells.
As outlined in dpeg's ranged combat overhaul, branding spells should no
longer be permanant, nor apply directly to the ammunition. This commit
removes all previous "<X> Ammunition" spells, including the original
"Poison Ammunition" spell.

Instead, branding ammunition is now done via the launcher and the
currently existent weapon branding spells. Currently, those that can be
applied to a launcher are: Fire Brand, Freezing Aura and Poison Weapon.
There is the capacity of Excruciating Wounds to also be applied, but
this would require the addition of the Pain ego for missiles, of which
the mechanics would need to be decided upon first.

Like temporary brands on weapons currently, these brands on launchers
are permanantly affixable by using a scroll of vorpalise weapon.

This commit also overhauls the Arcane Marksman class. It began by
removing the non-existent ammunition spells; this left Devastating
Missiles with two spells: haste and deflect missiles. On reflection, I
decided to delete the book instead.

The Book of Elemental Missiles was easily adjusted by replacing the
three branding spells by the existent (and aforementioned) branding
spells. As there are no (according to the 0.6 ranged combat overhaul
spec) spells intended for use with translocation effects on missiles,
this left The Book of Warped Missiles rather empty.

Attempts were made (with discussion with Eronarn on IRC) to work out
something else to fill up the Book of Warped Missiles with, but I made
the decision to get rid it instead, as the only good suggestions
basically made it the Book of Spatial Translocations, and in that case,
the player may as well select the warper class.

We also discussed whether or not to remove Poison Weapon from the
Elemental starting book; I decided it was best to leave it now, but
increase its level from 2 to 3, to give the book more of a range of
levels (starting a Spriggan Arcane Marksman had all the spells listed as
"excellent" or "very good"; more on this shortly). I also decided that
Repel Missiles is over-used in this context, and upon Eronarn's
suggestion and sorear's agreement, have replaced it with Cause Fear.

I think this will provide for interesting game play, and that the level
of the spell is suitably high enough for it to not be abused. I also
made the executive decision that "Elemental Missiles" was no longer an
appropriate title, as it didn't contain any ammunition-specific spells
any longer, and have renamed the book to the "Book of Brands".

It is now:

Book of Brands

Level 1, Corona
Level 2, Swiftness
Level 2, Fire Brand
Level 2, Freezing Aura
Level 3, Poison Weapon
Level 5, Cause Fear

As the Arcane Marksmen class is now launcher-specific, rather than
ammunition-specific, I have adjusted it to no longer give Hill Orcs,
Sludge Elves and Merfolk javelins and throwing nets (they get bows like
everyone else), and made the decision to ban Trolls and Ogres from the
class altogether.

The reasoning here is that they have bad launcher aptitudes, and as they
are described as "marksmen", it makes no sense to give them a book full
of launcher-specific spells and some large rocks and throwing nets. This
change could, of course, be reverted, and they could simply be given
bows, crossbows or slings as relevant, in line with all other races
being given launchers and ammunition.

I have adjusted the starting skill levels for enchantments, spell-
casting and dodging to that of Crusaders. I have also removed Poison
Ammunition from the Book of Envenomations, and replaced it with Poison
Weapon instead.

Finally, I have increased TAG_MAJOR_VERSION to 14.