Another ranged combat target: branded launcher + branded ammo.
This specifically allows for the combination of poison + flaming arrows,
so that Nessos is not unfairly nerfed. Otherwise, launcher brands will
not affect ammunition at all. They will only apply to unbranded
ammunition.

I'm not too sure on this myself. The idea that a flaming bow does not
affect a missile of frost, or vice versa, a freezing bow does not affect
a missile of flame is a little bit odd. It's possible that I
misunderstood, and if so, this commit should be reverted.