Fix MP+X property in unrands (Mantis #294, Vandal).
calc_mp() only seemed to be called when games are loaded, meaning that
putting on the unrand Hat of Pondering did not cause the player's max MP
to immediately increase.

Saving and reloading fixed this. The same happened with removing the
hat: the player's max MP stays at the same value as when the hat was
worn. Saving and reloading also fixed this.

Fixed this by calling calc_mp(), and also cleaned up the code slightly.
It still doesn't seem to be printing the "Your magic
increases/decreases" messages, however.