Rods have an intrinsic recharge rate (#300)
Using Eronarn's proposed system.  Rods get a baseline of 4 MP per 100
turns; every rod has an enchantment, which adds to this.  On top of
that, you get the old Evocations bonus.  At maxed Evocations with a
+9 rod, you gain 52 MP per 100 turns.

Unfortunately, all four of the multi-purpose fields were already used
for rods, so I had to spill enchantment into props.