Implement leading dash.
This is now output for each new turn, not after each mesclr() as
used to be the case with delay_message_clear.

The dash is printed every time the turn count increases from one
message to the next. This sounds right, but leads to weird
messaging because monsters appear to act first on a turn.

Also, the circular buffer no longer tracks its size.