Don't crashdump on SIGTTIN and SIGTTOU
Neither SIGTTIN nor SIGTTOU is a fatal signal; the default handling
for them is fine for Crawl (suspending the process until a tty is
available for the input/output it needs), but before this change it
causes an (incorrect) crashdump. This change simply removes the
crashdump in this case, as there isn't actually a crash.

Signed-off-by: Stefan O'Rear <stefanor@cox.net>