Saving throws for new needle brands (syllogism, Twinge, others).
The current needle brands are overpowered in the fact that you can
practically paralyse or confuse anyone (including Cerebov) with little
or no effort.

Twinge has kindly provided a workable formula: it includes a 2% chance
of automatically applying the brand to monsters with less than 15 HD.
The raw formula is:

2 + random2(4 + throwing skill + blowgun to_hit) > monster HD.

Twinge also provided the following statistics of success: With no skill
you're still at 100% to proc on HD1 or HD2 (gnoll, D1 monsters), 75% on
HD3 (Sigmund), 50% on HD4 (Divessa), and 25% on HD5 (ogre). With say 4
skills as an early Assassin, you're up to 56% on an Ogre, or 33% on a
Yak.

We still may need to investigate the possibility of resistances; with
this code, there is still a chance of confusing or paralysing Cerebov,
but one would need a high Throwing skill and an extensively enchant
blowgun to achieve.

Eronarn has suggested re-using the multiple levels of resist poison, or
possibly using a new type of multi-level resistance to "organic
compounds".

Finally, descriptions will need to be updated somewhere to include
information about high skills and enchantments allowing you to overcome
monsters' resistances; this currently does not include poison or curare,
though.