Call them "<missile> of <brand>", not "bolt of <brand>".
Using the bolt terminology provides confusion if the weapon in question
is a sling or a bow and not a crossbow, and as the relevant brands are
"frost" and "flame", it causes confusion with "bolt of cold" and "bolt
of fire".

The relevant bolt names for wands of flame and frost are "puff of flame"
and "puff of frost".