New type of cloud: thick gloom.
It's a fast-moving, long-lasting cloud that clumps together. It will
dissipate very quickly unless it is near other clouds. It is opaque, and
blocks line of sight with two or more clouds, but unlike other clouds,
it does not actually use the '#' glyph. Instead, it colours the ground
underneath it magenta.

It's currently unused, and while it's an interesting visual effect that
will have some great potential in portal vaults and specific branches,
should probably be used very rarely.

Also includes a tile for these, though I'm not too happy with the
effect.