Blink quietly when timing out monster confusion.
monsters::timeout_enchantments is called when returning to a
level; it blinks monsters that were confused and used to message
that the monster blinks. This fixes that message.

Things are still a little weird: You still see the confused monsters
when entering the level and get the "less confused" messages before
they're redrawn elsewhere. Not quite sure how this should be done
properly.