In trog_book, burn all books at the book location.
Previously, it was possible to prevent the book from
being burned by throwing an item on top.

The timing of incineration has been a little unreliable in
in my tests, but I believe (hope) this is not related to the
change.

Burning all books except just one so players can't throw
away a junk book to get at a new one.