Make slime creatures split in open space.
I removed an "unoccupied" because a comment claimed this shouldn't
matter for splitting and also because doing so seemed to have the
desired effect.

I also added a check to make sure that splitting only takes place
if it doesn't increase the distance to the foe, other adjacent
cells are fine of course.

With these changes, luring merged slime creatures into open areas
causes them to split.