Fix incorrect math in make_box_doors.
In most settings, three passable glyphs are those on the inside of the
room; we want some passable glyphs on the *outside* of the room as well.
This does mean that immediate corners of doors will now be excluded from
possible door spots, but this can be hacked around. The other solution
is to look "outside" of the room, which could probably be calculated
from the specified side.