Add new props value "tile_num" and use it for randomized monster tiles.
Monsters that have several variant tiles will now get a random value
written into tile_num at creation, so when their tile gets drawn one
of the set is picked, but always the same one.

In practise, this is currently only used for slaves and toadstools.
Other monsters that have variant tiles, such as ugly things, kraken
tentacles, or slimes creatures, are handled specially.

Generally, such differences should be small and flavourful. The greater
the threat of a given monster the more important it is that the player
will recognize it immediately. (Of course, this restriction does not
hold if the tile actually transport information as is the case for
ugly things' colours etc.)