New method _command_to_string() returning human-readable key names.
The plan is to replace the hardcoded command key names on the help screen
and in the tutorial with these dynamic ones respecting redefined keys.

Unfortunately, several commands have multiple assignments and macro.cc's
command_to_key() only returns the last one, which might not be what the
player is actually using. For example, I don't have a Numpad, so all
mentions of it are of no use to me, and I'd rather have the vi keys
listed, which in turn many other players have no interest in.

You can check the current key assignments if you comment out the
DEBUG_DUMP_COMMANDS definition in chardump.cc and then create a character
dump. This will be removed once I'm finished here.