More dynamic command key listings on the help screen.
Except for the movement keys (uh, help?) and the item commands that have
their starting letters highlighted (again, help?) all commands are now
shown as actually assigned.

As I said before, the key that's get displayed might not be the one the
player actually uses. Likewise, there are still a few instances where
alternative keys get listed that might no longer be correct and have been
assigned elsewhere.

I'm sorry to see that the command help is no longer WYSIWYG[*] for
developers (so recompilation cannot be avoided), but it _is_ WYSIWYG
for players, which is much more important.

The best way to test this is to include the new dvorak settings and
have a look at the ?? help screen.

*) What You See Is What You Get