Convert Swamp ending vaults to work with new Swamp layout.
This basically converts all instances of rock to trees. It looks quite
good, and I've kept the stone as thematically, this works in the branch.
This commit also increases the chance of an azure jelly in the Ice
ending, instead of an equal chance of an azure jelly or an ice dragon.