New Swamp ending: "old school" Swamp, Lernaean hydra.
This ending reproduces the look and feel of the "old" Swamp: this
specifically includes weird corridors through passages of water,
randomly placed rock walls, and randomly placed deep and shallow water.

It is considerably larger than most other ending vaults in order to
continue this theme. The map has a similar number of hydrae and other
threat to the original swamp and swamp_alternate maps.

As a bonus, this map contains the Lernaean hydra! It's tucked away in a
corner. Blame greensnark for this. It can be removed if it proves too
much of a killer.