Tile transparency in water without overlays.
In order for jpeg's waves to work on deep water as well as shallow, we
need to not use partially transparent overlays to simulate an actor or an
item being submerged.  Now, non-flying objects on water will be drawn
transparently to blend with the water below.  This should mostly look the
same, except it will now work on top of waves and will not require a mask
for each water type.

As a nice side-effect, ghosts are now transparent again and the water on
top of submerged objects now animates properly.

See the comments in tilebuf.cc for details.  The mask tile itself can be
adjusted to change the water level, but the parameters to
SubmergedTileBuffer will need to be changed to compensate depending on
what the new art looks like.