Fix allies making a beeline for the map origin when out of LOS
Fix a bug introduced in commit 8e49c668 which caused the movement
target of friendly monsters outside the players LOS to be set to (0,0).

monsters::get_foe() returns NULL for friendly monsters with foe equal
to MHITYOU, so it was not being used correctly in handle_behaviour
(since this part of it is just concerned with where monsters are trying
to move rather than what they will actually attack).