Fix statues in dis_mu not coming to life (Mu_).
Kind of a hack, in that it's just reverting to the previous version of
hells.des. It seems like the monster/slave integration doesn't work
properly in this regard. The best I could do was to get Dispater's death
to convert one statue (instead of making a statue where he died).

This now converts all statues.