New events: door opened, closed. Lua marker for door restriction.
The events are triggered whenever a door is opened or closed by the
player, or opened by a monster (monsters can never close doors). These
events can be used by Triggerables.

The RestrictDoor Lua marker restricts monsters from opening doors unless
the player has already opened them. In combination with teleport
restriction, this allows wandering monsters to not interfere with loot
chambers. Obviously, the uses of this should be limited.