Allow monsters to cut through plants.
Fixes #221 (and the duplicate #408).

Some people think that this used to work in the past, but I couldn't
find traces of this in the code. 0.5.2 at least doesn't appear to
have monsters cutting through plants.

This works by letting monsters attack their mmov target if it's
a monster with mons_is_firewood and there are no good moves, so
that monsters will walk around plants if possible.