0000460 : "you hear a splashing noise." message leak
Modified player_can_hear, added hear_distance argument which defaults to level-wide (so as to not break any previous uses of player_can_hear which represents the distance a player can hear). Modified monster.cc to pass in a parameter of 20 to player_can_hear when generating the splashing message.