More prompts for full window (in case of clear_messages=true).
There's also some code to do more prompts in the delayed case,
but that is incomplete (would maybe prompt for the next
new-turn message). That part is designed to save a line by
using the bottom left character as a prompt -- that might
be an option for clear_messages also, perhaps with a
coloured '+' in the bottom right corner.