Make line_reader aware of its region.
This should provide more clarity around the wherex()/wherey()/
get_cursor_region() confusion, which has to do with wherex()/wherey()
being screen relative in console mode and region relative in tiles.

It's still a hack, thought, and untested in tiles at the moment.

The main reason behind this is to eventually allow the line_reader
to scroll the region it's working in (see #484).