Move Tomb maps into their own files.
Apart from the Hall of Blades, which is just a single map in vaults.des,
that is every branch with its own map file. I think this is much
cleaner, and means that we don't have to scroll through all the
different branches and branch ends just to find the one being looked
for.