Allow line_reader to scroll its region.
This fixes issue #484.

There's still the problem that it will wrap to screen column
zero -- a region-aware version of wrapcprintf would be
required to fix this.

It's currently limited to the message window, since that's the
only region that implements scrolling. line_reader itself takes
care of printing its scrolled buffer.

This is something of a hack. It might eventually be desirable
to have some more window-like regions that all store their
content and allow scrolling; there would be some overlap with
tiles' TextRegion etc. But that would mean redoing curses
windows on top of cprintf/cgotoxy...