Cheibriados piety loss for unequipping ponderous armour.
This solves the problem that before, players were urged to
wear their ponderous armour against easy opponents but not
when exploring and against strong opponents.

Taking off ponderous armour cuts piety by 1/4.

It would be nice to allow the player to switch armours
while retaining the same level of ponderousness, but it's
unclear how to do that correctly and elegantly. Suggestions
welcome (Chei wiki page).

If this is done, the piety cost could be raised by a bit.