Kraken no longer submerge while fleeing, but squirt clouds of ink.
New cloud type of "ink": it can only exist "over" water (and should be
detailed in Tiles transparently, as though it's under the water as well
as above it, if posible).

Also, tentacles will now flee when the kraken head flees. This means
that the kraken head doesn't flee, and then stop because the tentacles
are too busy attack you to do anything interesting. It does't mean it
will always continue fleeing, however.