Don't attempt to unpack non-existing .tdl files.
A character created in the console version naturally doesn't have a
tile doll (tdl) file and if you switch to the Tiles version you
kept getting error messages about said file not being found until
you eventually open the savefile in the Tiles version. This is
no longer the case.