Convert items Lua library to use metatable. May case end-user issues.
Like the monster library, the item library now uses a metatable to store
its functions. Instead of using lightuserdata, as before, and passing
this to the different item.* functions, the functions are now members of
the actual item.

The item library has also been renamed to "items": this allows for its
use in Vault definitions without it clashing with the item specifier and
dgn.item function.

Example:

 * Previously, you would use:
     local x = item.inventory()[1]
     print(item.name(x, false))
 * Now, you would use:
     local x = items.inventory()[1]
     print(x.name(false))

This will have an impact on end-user scripts that use the old version of
the item library. I have already updated the uses of the items library
in the dat/lua scripts, as well as vault definitions and marker
definitions. I think I've caught them all.

There may be bugs or issues, but hopefully that's everything. (Does
cause issues with annotations such as {artefact}, will be fixed
shortly.)