Don't hand out darkgrey to 'BLACK' monsters. (TGW)
Monsters that specify BLACK are given a random colour on placement.
Unfortunately, due to the fact that random_colour() picks any colour,
this means that these monsters sometimes end up with DARKGREY, which
clashes with our policy of not using DARKGREY as a monster colour.

This commit introduces random_monster_colour(), which will re-pick a
colour if DARKGREY is given, and continue to re-pick until it gets a
colour that isn't DARKGREY.