Tweak slime_altar_2 to be less instantaneous death, but still threatening.
Unfortunately, as there's no indication as to what's behind the door,
and there is quite a high chance of acid blobs (I got two in one game),
the moment you open the door, you're pretty much dead unless you have a
scroll of blinking or some instantaneous means of escape.

I've therefore tweaked the vault and placed glass around the door, so
that you can see inside, and also placed a warning and a restriction on
the door: the first is so that you don't auto-explore it open, and the
second is so that the jellies don't see you and decided to come out and
visit.

It still presents an interesting threat, but not one that you are forced
to deal with immediately.