Hopefully finally fix "of flame of flame" and other errata.
This was not helped by the fact that ammo_name, as passed to
setup_missile_beam, is only used by the player throwing functionality.
Flaming, frost, and chaos missiles will now identify when thrown, so you
shouldn't see "XYZ shoots an arrow of flame" and then pickup "a glowing
arrow".

You will still get "XYZ shoots a glowing arrow" for some types of
missile (dispersal, etc), but these should identify as per normal.

This hopefully sorts everything out, but really should only be
considered a temporary fix: the code is messy and confusing as hell, and
needs to be rewritten to be clearer, and to use the same naming and beam
structure code for both.

This commit doesn't address the issue that casting poison weapon or fire
brand doesn't identify on the weapon, something I noticed occuring when
using &I in WizMode. Unsure if it's actually an issue in non-WizMode
play, so I'll look into it.