Nerf poison arrow.
Halve the irresistable part from 60% to 30%.

Poison arrow has seen a significant relative boost
with the range nerfs to other conjurations.

As compared with other options (raise level, lower range),
this has the advantage of furthering differentiation between
schools: It is not desirable that a conjurer should want to
branch out into poison magic to handle the omni-resistant
end-game monsters.

If this turns out to be too hard on the spell, a more
interesting buff than higher damage would be to make
the spell quieter to cast -- currently, only non-conjuration
poison magic gets reduced spellcasting noise.