Hack monster LOS to reduce updates.
It's now only updated if the monster has moved or is in view of
the player. That means it will not necessarily be accurate if
terrain or clouds are LOS-affecting monsters change/move.

I'm not sure that this is much better than just not updating
LOS for sleeping monsters -- depends on to what extent monsters
just stay in place.

Previously, crawl spent a significant amount of time with the
monster LOS updates.

Other options:
- Calculate monster LOS on-demand (unlikely to help much)
- Reduce monster LOS range away from the player
  (this might require some changes to LOS code to actually save
   time)
- Track LOS globally: for each cell, which other cells are visible?
  An estimated 200 kB of data.
  This would then only be updated if terrain/clouds/LOS-affecting
  monsters changed.