Move giant spores to '*', don't spawn them in Slime.
Giant spores used to spawn in Slime, assumedly because the 'G' made them
a type of eyeball/slime. However, with recent changes to ballistomycete
spawn mechanisms and the introduction of Fedhas, spores are definitely
classed as a type of plant, so their placement in Slime is now dubious.

Moving them to the '*' glyph also brings them in line with other
monsters that "explode": specifically, ball of lightning and orb of
destruction.