More glyph changes: move elementals to 'i'.
Quite a dramatic change, but it fixes the last of the issues that exist
with glyphs: monsters sharing the same glyph as features is confusing
and problematic.

This commit makes the following changes:

 * water elemental lightblue "{" becomes lightblue "i"
 * earth elemental brown "#" becomes yellow "i"
 * fire elemental yellow "#" becomes lightred "i"
 * air elemental white "v" becomes white "i"

The "i" glyph was previously unused for anything. I've also tweaked the
colouring so that they are all lightly coloured, made the fire
elementals lightred instead of yellow, and gave earth elementals yellow.

Hopefully, a likewise non-contentious issue. "i" may change at a later
date, but I think that having the elementals off the feature glyphs is
the most important part of this change.

Previous suggestions were to move efreets to "R" and put elementals on
"E"; this could still be done, but it would cause colour confusion
between Mara and Azrael, and also requires shifting of other monsters,
while this change doesn't require any further shifting.