Lua wrapper for travel_stoppers, auto-add portal vault messages.
Hits one of the items on jpeg's wish-list (hopefully!). This commit adds
a Lua (clua-accessible) wrapper for adding phrases to the travel stopers
list. It does't currently deal with message channel travel stoppers at
the minute, and just uses the string as-is (this can be changed).

This commit also adds functionality to lm_tmsg.lua that automatically
includes any portal vault messages (with entities included) in the
travel stoppers list.

Hopefully, it works properly. I haven't been able to figure out a sane
way to test it yet.