Use KMASK no_item_gen to prevent Triggerable issues.
As experienced by Napkin on CDO, Triggerable's using the "auto" target
for items complain if there is more than one item at the location in
question; as this usually involves linking to the Orb (or rune) and
teleport control, it can be problematic.

Tagging the runes/Orb with no_item_gen should mean that random items
aren't generated on top of the rune/Orb, and therefore should prevent
this from occuring.

Temporary fix, though, as there's probably a way to improve the
Triggerable identification method.