Experimentally overlay demon tiles with the corresponding numbers.
I added an option tile_show_demon_numbers, defaulting to true (at
least for testing purposes). Not documented yet, as things are still
subject to change.

One of the big disadvantages of tiles vs. console is that it is
impossible to get an idea of demons' difficulty from their tiles.
Tiles players have to learn the connection between their tiles and
danger (and it's very easy to get them mixed up, probably more so
even than their names), whereas console players can tell at a glance
that a '1' is more dangerous than a '3'. These numbers now also get
displayed for tiles, which is heavy-handed, technical and confusing
for players otherwise unused to arbitrary glyphs. ("Why is 1 worse
than 5?"). I've also coloured the numbers in a range from yellow (5)
to red (1), which is much easier to understand, and this at least
should be kept even if the numbers get replaced with other symbols.

For example, if we interpret the demon glyphs correspond with their
"ranks" in hell, the tiles could display stars instead, or skulls,
or coloured flames. Whatever, anything but numbers. :p