Add icons for armour brands.
I'm still not sure myself whether this is actually a good idea. On the
one hand we display information that is already available but currently
requires more "work" (hovering the mouse over the item) to see, on the
other hand you don't change armour remotely as often as you do jewellery
or even weapons, so it might be information that is not really necessary.

If we do keep them, the icons (mostly copied together from potions,
jewellery etc.) need to become smaller. I think that both my own archery
tile and Porkchop's reflection one have a good size, not as intrusive as
some of the resistances covering a quarter of the armour tile but there
if you look for it. Also, silence and archmagi only have placeholder
tiles at the moment.