Make monsters lose energy when noticing the player.
This is EUT_NOTICE, which currently defaults to 4. This should mean
that monsters are unlikely to get double actions on the turn they
notice you.

A lower value than 4 might well suffice. A higher value might be
necessary if we want to avoid centaurs double-firing when you move into
their view.