Rename "giant iguana" to "iguana".
As TGW points out, green iguana grow to an average of three to six feet
long; gila monsters only grow to a maximum of two feet long, but are
defined as being larger and heavier than iguanas.

I think that using "giant" as a metaphor for "dangerous" isn't
particularly necessary, threats can be identified in other ways.