Fix "was cursed was cursed was cursed" ad infinitum.
Autoinscribe cursed will no longer add a "was cursed" inscription if the
item was previously cursed. There have been some suggestions that only
items which you didn't see being cursed should be marked as "was
cursed".

This would require some sort of tracking flag, ISFLAG_SAW_BEING_CURSED,
or something similar to this, to implement properly. A bit too much work
for now, so this should suffice as a simple fix.