Make Che support the use of ponderous armour per dpeg's design.
Che supports up to piety-rank (number of *'s) pieces of
ponderous armour, giving resists rN, rC, rF for the first three,
and +1 to all stats for each.

The code mostly lives in the new file godpassive.cc that's meant
to collect passive god effects. Changes to piety and number of
equipped ponderous items result in calls to che_change_level
which will do what needs to be done.