Tweak Mara resists, experience levels.
His experience was reduced by reducing his HD, so I've increased the
modifier to keep it in the 7000 experience points range (to compensate
for the fact that he is roughly three uniques put together).

I've also removed the electricity resistance and instated two levels of
fire resistance, to compensate for the fact that the clones will shoot
through each other.