Framework for portal vault entries having tiles; Volcano tile (purge).
Unfortunately, just using tile("O = <xyz>") is not enough, as the tile
won't go away when the portal is actually removed; I've added a new Lua
function "tile_feat_changed", which specifically alters
env.tile_flv(gc).feat, allowing you to change the override feature tile
on-the-fly.

Probably could do with tile_wall_changed and tile_floor_changed
variants, but they aren't needed at the minute and I'm not sure I can
see a use for them just yet.

I've also altered the 1way stair (and timed stair marker variants) to
accept a "floor_tile" as well as a "floor" parameter. If provided, this
will be the string value that is passed to tile_feat_changed when the
marker times out (or the player uses the stair).

If not provided, it will default to no tile (using the default tile of
whatever feature is provided in "floor").

Finally, use Purge's dark tunnel tile for Volcano entry. Thank you!