Add a check for feat==floor in feature's tile override.
This means that a feature that is converted to a floor tile (such as an
archway when timing out portal vaults, or a statue when disintegrated)
will revert to use the floor's tile (or the default for that level)
instead of the overridden tile being displayed still.

I've left the code that changes the tile in the timed markers (and 1-way
stairs) as people likely won't want to use the same tile for the
timed-out feature ("empty arch", or whatever) as they would for the
initial entrance.