"Mirror" wall tile, use on Tukima.
A slightly tweaked version of tile by purge that was based on a tile I
was making that was based on the green crystal tile! This one is
slightly "whiter" and more "crystal clear" than purge's edit, and I'm
still not sure if it's perfect.

It looks pretty neat in-game though, and I'm over-all happy with the
effect. Colour seems roughly equivalent to the "silver" altars.