[470] Fix opaque white shallow water in tiles.
Alpha testing is apparently not a dependable feature of OpenGL.  Instead,
provide an alternate method of drawing submerged tiles that does not
depend on it.  The older (better looking, in my opinion) behavior can be
enabled via an option.  It's off by default, so that folks don't file bug
reports about shallow water.