Use the circles version of Fedhas' oklob retribution more often
Make Fedhas use the concentric circles version of oklob retribution
more often (relative to the single circle version, the chance of
getting oklobs relative to other forms of retribution is unchanged).

This commit allows the oklob circle to be placed as low as radius 4,
previously it could only be placed at radius 6.