Annotate exclusions based on map knowledge.
This means proper annotations for remembered monsters.

For detected monsters, append a "(detected)" to the description,
which probably suffices with the current state of divinations.

Note that env.map_knowledge will contain the detected base monster
(mons_detected_base), so putting an exclusion on detected Anteaus
will currently result in "ettin (detected)", or
"small zombie (detected)" on a curse skull. At some point, we
could add
  std::string mons_detected_class_name(monster_type)
that would resolve MONS_ETTIN to "giant" if desired.